Breaking the graphql assets
